@@ -153,7 +153,7 @@
         "markdown-translatable": "^1.3.0",
         "md5": "^2.3.0",
         "moment": "^2.29.3",
-        "next": "^13.0.0",
+        "next": "^13.1.6",
         "next-images": "^1.8.4",
         "next-on-netlify": "^3.0.1",
         "next-optimized-images": "^2.6.2",
