@@ -1,9 +1,9 @@
 import Signup from '../src/components/Signup/Signup';
-import Meta from '../src/Meta';
+// import Meta from '../src/Meta';
 
 const signup = () => (
   <div>
-    <Meta />
+    {/* <Meta /> */}
     <Signup />
   </div>
 );
